test(orchestration): split dispatch-liveness tests out of db.test.ts
db.test.ts sits just under the 800-line max-lines budget, leaving no
room for new coverage. Move the recordHeartbeat and getStaleDispatches
tests into a new db-dispatch-liveness.test.ts — a coherent unit around
detecting whether a dispatched task is still alive — and leave the
dispatch-context CRUD coverage behind, which is a different concern.

No assertions or fixture behaviour changed; only file location.
